add tests for data_sources_outputs
