+++ tests/data_sources/test_datasource_output.py
@@ -1,75 +1,62 @@
-""" Tests for data_sources """
-from nowcasting_dataset.data_sources.fake.batch import (
-    gsp_fake,
-    nwp_fake,
-    pv_fake,
-    satellite_fake,
-    sun_fake,
-    topographic_fake,
-)
+"""
+Test for datasource output validations
+"""
+import numpy as np
+import pytest
+
+from nowcasting_dataset.data_sources.fake.batch import gsp_fake
+from nowcasting_dataset.data_sources.gsp.gsp_model import GSP
 
 
-def test_gsp(configuration):
-    """Test gsp fake"""
+def test_datasource_output_validation(configuration):  # noqa: D103
 
-    configuration.process.batch_size = 4
-    configuration.input_data.gsp.history_minutes = 2 * 60
-    configuration.input_data.gsp.forecast_minutes = 4 * 60
-    configuration.input_data.gsp.n_gsp_per_example = 32
+    configuration.process.batch_size = 2
+    configuration.input_data.gsp.history_minutes = 60
+    configuration.input_data.gsp.forecast_minutes = 60
+    configuration.input_data.gsp.n_gsp_per_example = 6
 
-    s = gsp_fake(configuration=configuration)
+    gsp = gsp_fake(configuration)
 
-    assert s.power_mw.shape == (4, 13, 32)
+    GSP.model_validation(gsp)
 
+    # nan error
+    gsp.power_mw[0, 0] = np.nan
+    with pytest.raises(Exception):
+        GSP.model_validation(gsp)
 
-def test_nwp(configuration):
-    """Test nwp fake"""
+    # inf error
+    gsp.power_mw[0, 0] = np.inf
+    with pytest.raises(Exception):
+        GSP.model_validation(gsp)
 
-    configuration.process.batch_size = 4
-    configuration.input_data.nwp.history_minutes = 0
-    configuration.input_data.nwp.forecast_minutes = 60
-    configuration.input_data.nwp.nwp_image_size_pixels = 64
-    configuration.input_data.nwp.nwp_channels = ["test_channel"] * 8
-
-    _ = nwp_fake(configuration=configuration)
-
-
-def test_pv(configuration):
-    """Test pv fake"""
-
-    configuration.process.batch_size = 4
-    configuration.input_data.pv.history_minutes = 30
-    configuration.input_data.pv.forecast_minutes = 30
-    configuration.input_data.pv.n_pv_systems_per_example = 128
-
-    _ = pv_fake(configuration)
-
-
-def test_satellite(configuration):
-    """Test satellite fake"""
-
-    configuration.process.batch_size = 4
-    configuration.input_data.satellite.history_minutes = 30
-    configuration.input_data.satellite.forecast_minutes = 30
-    configuration.input_data.satellite.satellite_image_size_pixels = 64
-    configuration.input_data.satellite.satellite_channels = ["test_channel"] * 8
-
-    s = satellite_fake(configuration=configuration)
-
-    assert s.x is not None
-
-
-def test_sun():
-    """Test sun fake"""
-    _ = sun_fake(
-        batch_size=4,
-        seq_length_5=13,
+    # fill nan
+    gsp.power_mw[0, 0] = np.nan
+    gsp.__setitem__(
+        "power_mw", gsp.check_nan_and_fill_warning(data=gsp.power_mw, variable_name="power_mw")
     )
+    assert gsp.power_mw[0, 0, 0].values == 0
 
+    # greater than
+    gsp.power_mw[0, 0] = 0
+    with pytest.raises(Exception):
+        gsp.check_dataset_greater_than_or_equal_to(
+            data=gsp.power_mw, min_value=1, variable_name="power_mw"
+        )
 
-def test_topo():
-    """Test topo fake"""
-    _ = topographic_fake(
-        batch_size=4,
-        image_size_pixels=64,
-    )
+    # less than
+    gsp.power_mw[0, 0] = 0
+    with pytest.raises(Exception):
+        gsp.check_dataset_less_than_or_equal_to(
+            data=gsp.power_mw, max_value=-1, variable_name="power_mw"
+        )
+
+    # equal to than
+    gsp.power_mw[0, 0] = 0
+    with pytest.raises(Exception):
+        gsp.check_dataset_less_than_or_equal_to(
+            data=gsp.power_mw, value=0, variable_name="power_mw"
+        )
+
+    # dims
+    with pytest.raises(Exception):
+        gsp.check_data_var_dim(data=gsp.power_mw, expected_dims="fake_dim")
